Start rolling back structured concurrency changes
I would have loved to get this working, but ran into a wall of
complexity when trying to make the spawner work fully. Joining all the
actors that the spawner spawned and getting all of their results seems
to require a fair amount of complexity to get working, and I'm not up
for it.

I'm sure it's possible, that this is a side show born out of a whim,
mostly, and I'm not going to sink any more time into it.