Switch number type to `u32`
`f64` is problematic, since it doesn't traits that I need.

Eventually, this shouldn't be a problem here, since `f64` is the wrong
type to represent a floating point _literal_ anyway. But I don't want to
come up with a future-proof design right now, and `u32` will do for a
while.