Remove erroneous sanity check
Turns out that hashes can be duplicated quite easily, and that is a
completely normal part of operations. That I thought that shouldn't
happen (and it still happening would be a bug), was based on a flawed
mental model based on previous prototypes that worked differently, and
where hashes were used for different purposes.

I'm going to add some documentation to clarify this in a follow-up
commit.