Explicitly shut down, if CTRL+C is received
I'm not actually sure that this code works in the current context. The
terminal is not in raw mode, and CTRL+C has been working the whole time
as expected, because of that. So the behavior is the same, with or
without this change.

I'm working on putting the terminal in raw mode though, and this code
will definitely be needed then. I have already tested it in that
context, and it works there.