Move `editor` module into `language`
This might seem a bit weird, but I'm going to reframe (and rename) the
`language` module as the core components that are independent of a
specific platform, or a specific host. And what's being moved here
definitely fits that bill, as all the platform-specific editor
functionality lives in `io::editor`.