Rename `core` to `lang`
`core` is turning out to be inconvenient, because the collisiong with
Rust's `core` library is confusing the IDE tooling.

I've been a bit squeamish about calling the module `lang`, because it
also contains the editor. But I think that's actually good. That name is
making the point, that the editor is just as essential to what this
language is about (immediacy) as something like the compiler is.