Add additional `identity` function test
This is the second test for the `identity` function, and like the first
one, it passes without that function existing yet. Again, due to the
rather primitive state of the language implementation.

But to satisfy that test, I had to advance the implementation
significantly, to something that slowly (_very_ slowly) starts to
assemble a real programming language implementation.

Together with the other `identity` test, this one is useful for
providing a baseline, as more functionality is added.