Extract `NodeKind` out of `Node`
This reverts an earlier design decision. What I intended by merging them
in the first place, being able to tailor the children to each variant,
hasn't borne fruit so far. But having the children per-variant has been
more complicated to deal with, and is getting in the way.

Maybe I'll merge them again later, if that turns out to be advantageous.
But it's clear to me now, that merging them earlier had been premature,
at the least, if not the wrong decision outright.