Prevent panic on latest version lookup
This panic isn't triggered by the test suite. I'm not actually sure if
it's possible to trigger, or if the workings of the rest of the codebase
somehow prevent this currently.

But whatever the case may be, I'm about to make changes that can trigger
it. So this addition is definitely necessary, and will be covered by the
test suite soon enough.