Update implementation note
The idea I had there before, using a `*const Node`, wouldn't work. Nodes
are stored by their `NodeHash`, so this would be equivalent to use the
parent's `NodeHash`, which is not enough to uniquely identify the
parent, for the same reasons that a `NodeHash` alone is not enough to
uniquely identify any node.