Remove stale doc comment
Turns out that adding the error I didn't want to add was pretty easy by
now. The code required for that will still be applicable, when I add
function parameters. And I expect to be able to re-use the same code for
branches, once functions can have multiple of those.