Store errors by `NodeHash`
Storing errors by `NodePath`, means they no longer apply, if the parent
changes, even if the erroneous node remains unchanged. This is obviously
wrong.

And right now, all errors result from a node and its children. So using
a `NodePath` in the first place is both buggy and overkill.

I'm not going to add a test for this bug. Maybe I should, I don't know.
But as I started writing it, it felt silly. The simplified API that is
going to result from this change, is making that error impossible in the
first place.