Remove redundant test
This test was added to prevent a specific mistake I made from happening
again. But the evaluator looked much different then, and it's going to
change more in the near futures, with `Node::Application` becoming a
thing.

So it seems that this test is no longer necessary. And in fact, it's
turning into a hurdle for further refactoring right now.