Remove support for removing nodes
I'm currently working on making the editor less free-form and more
restricted. One of the consequences of that, is that removing nodes in
its current form won't be possible in every case, in the future.

Instead of slowly refactoring the current node removal functionality,
I've decided to just remove the feature outright. The editor isn't ready
for productive use anyway, and for testing, removing nodes is not
required.

This is going to make the coming transition easier. Once that transition
is complete, this feature can be re-added, in its new form.