Remove redundant test
This was to be the seed of a test suite for the replacement algorithm,
to support and enable the complexity that was going to be required to
prepare the replacement algorithm for future requirements.

But restrictred structural editing actually makes things a lot simpler,
so that algorithm can get simpler rather than more complicated. So not
only is that test suite not going to be needed, the existing test is in
the way of that refactoring.