Remove `SiblingIndex` field from `Parent`
Having the `SiblingIndex` field in `Parent`, means one `Parent` needs to
exist per child. Without it, one needs to exist per node that is a
parent, which creates a 1:1 mapping.

This is relevant for the refactor of `NodePath`, for which I probably
need a math of `Parent` to `NodePath` in `Codebase`.