Add `RuntimeNode::Apply`
I don't know yet, if this is a good direction. It's more code overall,
although about the same number of lines in `Evaluator`.

The trade-off here is that `Evaluator` gets to deal with nodes in a more
type-safe manner, while `RuntimeNode` becomes more complex. And I hope
that this will also allow plugging some holes in `Evaluator` related to
the call stack, once this is a bit more fleshed out.

For now, I'll continue along this route. We'll see if I have to
backtrack.