Remove game engine test suite
I'm currently working on restructuring how the game engine works, and
this test suite is getting in the way. As part of my work, I've decided
to give the game control over the renderer, which makes the whole system
inherently less testable. This test suite can't exist then.

It would be very desirable to make the system testable again in the
future, but I don't know what that should look like right now. I don't
think the current approach would scale to more complex renderer
interactions that need to be possible in the future. I think it's
probably best to see how that develops and re-evaluate later.

In addition, while it's still a shame to lose this highest-level test
suite, I think I've rarely ever seen these tests fail during
development, if ever. So maybe the lower-level test suites are actually
pretty good, and this loss is hopefully limited.